Moved timeseries and distribution plot into own files. Also refactored to use inferenceData
