+++ covid19_inference/plot/__init__.py
@@ -1,1 +1,3 @@
 from .distribution import distribution, _distribution, _plot_prior, _plot_posterior
+from .timeseries import timeseries_overview
+from .utils import get_array_from_idata, get_array_from_idata_via_date, format_date_xticks, format_k, add_watermark
